refactor(mcp): extract lifecycle logic into CompassMcpServerManager
Move all MCP server lifecycle logic (token generation, start/stop,
IPC handlers, preference listener) out of application.ts and into
CompassMcpServerManager in the compass-mcp-server package.

The call site in application.ts now mirrors setupCompassAuthService():
a 5-line method that calls CompassMcpServerManager.init() and registers
an exit handler via onExit().
